refactor: restructure database seeding process for improved maintainability
